Make most perl script have a common style
We had a mixture of perl styles, some of it being 8-column indent
Whitesmiths style, a little bit of GNU style (i.e. 2-column indent),
and the majority close to the perl style as described in 'man perlstyle'.

The work done in this change is to remove tabs, change the Whitesmiths
style brace position to the default perl style, adjust indentation to
4 columns where needed, and cleaning up some oddities left by earlier
'perltidy' runs, most notably this construct:

    }
    else {