Increase minimal Windows target version from XP to Vista
Support for Windows XP ended long time ago in April 2014.
Leaving Windows XP as default target version prevents
BCryptGenRandom() from being used when seeding the CSPRNG.

Although Windows Vista support ended as well in April 2017,
support for it is not dropped yet for the sake of compatibility,
because the distinction currently does not make a difference
in the code.

Note that it is still possible to build OpenSSL for Windows XP
(if necessary) by adding the -D_WIN32_WINNT=0x501 Configure argument.