Make SSL_OP_ALLOW_NO_DHE_KEX actually skip ECDHE
Currently, SSL_OP_ALLOW_NO_DHE_KEX is ineffective, as on the client side a key
share extension is sent regardless, and on the server side we will always send
a key share if the client sent a key share. What's more, the server will fail
with an internal error if a client key is missing later on in the exchange.

To fix this, we change the client side logic to skip sending a key share
extension if SSL_OP_ALLOW_NO_DHE_KEX is provided, a PSK session is available,
and we are not performing a hello retry. In order for this to work, the PSK
session is now determined when we generate the PSK KEX modes extension, which
comes before the key share extension in the handshake. We also fix the
server-side sanity check to allow a missing ckey if SSL_OP_ALLOW_NO_DHE_KEX is
set, and a PSK session is in use.