asn1parse: avoid double free
|str| was used for multiple conflicting purposes.  When '-strictpem',
it's used to uniquely hold a reference to the loaded payload.
However, when using '-strparse', |str| was re-used to hold the
position from where to start parsing.

So when '-strparse' and '-strictpem' are used together, |str| ends up
pointing into data pointed at by |at|, and is yet being freed, with
the result that the payload it held a reference to becomes a memory
leak, and there's a double free conflict when both |str| and |at| are
being freed.

The situation is resolved by always having |buf| hold the pointer to
the file data, and always and only use |str| to hold the position to
start parsing from.  Now, we only need to free |buf| properly and not
|str|.

Fixes #8752