Refactor the ex_data initialization to allow multiple inits
With multiple library contexts, it's possible that the same module
runs its init and teardown functions multiple times during its
lifetime.  This would normally not cause much trouble, except that our
FIPS module is a bit special and will include code that handles a
library context internally, just to be able to make internal EVP
calls.

This means that the provider init function will potentially be called
multiple times, and the teardown as well, and those may do
initializations and deinitializations, including the ex_data cleanup.
This may mean that the ex_data lock gets freed too early (on first
cleanup instead of last).

To ensure that the ex_data lock isn't freed before its last use, we
associate a reference counter with it, which gets incremented for
everytime there's an attempt to create that lock, and gets decremented
when there's an attempt to clean it up, thereby ensuring that
construction will only happen on first init and destruction on last
deinit.

Related to #8728