WIP: Add support for Intel CET
Intel CET adds new instructions to mark branch targets and is used to
defeat Return Oriented Programming (ROP) techniques. When CET is enabled
all indirect jumps or calls must end on an ENDBR instruction to signify
this was a valid jump target. If the instruction is missing a fault is
raised and generally the kernel terminates the program.

ENDBR instructions are NOPs on older CPUs that do not support them so
they can be added safely to without conditional compilation options.