Avoid zero-length memdup in tls_construct_ctos_early_data
The old-style PSK client callback is documented to deal in
NUL-terminated identifiers, and indeed we use strlen() to determine
the length of the identifier in question.  Since we are using a
memdup-family function rather than a strdup-family function, we must
accordingly increment the length being copied to account for the NUL
terminator.

Reported by M-Peter-Fth in github Issue #8894.