Configure: add -fkeep-inline-functions to --strict-warnings
The reason for doing this is that there are compilers (such as SUNPRO
CC) that do this by default, and this has consequences when header
files like safestack.h and lhash.h are included (possibly via other
header files) but the resulting program isn't linked with libcrypto.

Incidently, we have a test program that gets safestack.h included and
doesn't link with libcrypto (it shouldn't need to, since all it does
is to verify that it builds without macro clash), and is thereby the
perfect test victim.