e_devcrypto: acommodate hmac across forks
When doing HMAC, it is often desirable to open a session, call update
with the inner key, and then use copies of the original session to
handle different messages.

This can happen across forks (openssh does this), which causes trouble
with open cryptodev sessions, as they are kept by the kernel, and both
processes will end up operating with the same session.

To alleviate this, we can keep a limited cache of input data, and defer
cryptodev initialization until we are actually calling the cryptodev
update.  Since the HMAC inner key update will reside in the cache, they
will be copied safely across a fork.

This helps with performance as well, as the cost of the ioctl call adds
latency; keeping a cache reduces the number of small-payload calls.

Signed-off-by: Eneas U de Queiroz <cote2004-github@yahoo.com>