Back out the deprecation of EVP_MD_CTX_ctrl
Commit d5e5e2ffa deprecated the use of EVP_MD_CTX_ctrl in favour of
EVP_MD_CTX_set_params. However this latter function requires the use of
a provider, but a provider cannot currently be used for DigestSign/
DigestVerify operations or if an ENGINE is involved. The code to do this
will come later but isn't there yet.

Therefore the deprecation was a little premature. There are places in
libssl that need to use the old function, so we back out the deprecation
until such time as we are able to put it back. The documentation that says
the function is deprecated has not been removed since backing ths out is
considered a temporary measure.