SSL: Pass the master key by explicit parameters
The parameter keys for passing the master key down to the algorithms
handling that was done in a way that mimics the legacy EVP_MD_CTX_ctrl()
function.  We now change that to use an explicit parameter key, and
let set_params() type functions translate that into legacy controls
where needed, under the hood.