Use OSSL_PARAM_OCTET_STRING for passing the ssl3 master secret
OSSL_PARAM_OCTET_PTR suggest the data is long lived, and doesn't
require that the actual data is copied.  It doesn't matter, however,
if this is true for the master secret or not, since the underlying
functions (sha1_ctrl and md5_sha1_ctrl) consume the data anyway,
thereby not needing it to be long lived.

We therefore switch to use OSSL_PARAM_OCTET_STRING instead, and simply
pass the data pointer down for the data to be consumed.