Unbend error initialization
Error initialization did two things at the same time:

- Create a lock
- Load the standard strings.

Among others, the initialization is called when setting an error
status.

Because other parts of libcrypto that are also called when
initializing might need to set an error, it created a fragile
situation that could potentially interlock.

This situation is resolved by splitting the initialization in two
parts:

1.  Create the lock, which is now done as part of the base init.
    Any setting of error depends on this lock.
2.  Load the standard strings, which is done on demand as before.
    Setting an error does not depend on these strings existing,
    only the final extraction of error as printable strings is.

This simplifies the initialization situation, and takes away the need
for the internal init option OPENSSL_INIT_BASE_ONLY.