Make setting up atexit handlers part of the BASE init
This happened implicitly when ERR_get_state() got called, as it had an
init call with OPENSSL_INIT_LOAD_CRYPTO_STRINGS, and since errors
could potentially be triggered at any time, the atexit setup was
usually done early anyway.

Now that we've removed the OPENSSL_INIT_LOAD_CRYPTO_STRINGS init from
ERR_get_state() (because it caused a nasty loop), the atexit setup is
suddenly not there when it should be.

The best solution is to make the atexit setup a BASE init, but still
ensure that the user can disable it with an earlier init with
OPENSSL_INIT_NO_ATEXIT.