Refactor the approach for recursive init calls
At any point, we can expect OPENSSL_init_crypto() to be called
recursively.  This works, as long as no locking happens in code that's
executed again in the recursion.  We already had a similar situation
with the ERR code, and solved it with a rather ugly special case.

Here, we change the approach by keeping track of what initializations
have already been started, and simply never try to start them again.
We do so by masking away accumulated previous opts from the current
ones in the OPENSSL_init_crypto call.

There's a consequence, that any init opt can only be used once, even
if that particular initialization ends with aan error.