Add -copy_extensions option to x509 utility
-copy_extensions option specifies how extensions from SOURCE --
certificate request (if -req option present) or certificate
(if -x509toreq option present) -- should be handled when transferred to
TARGET -- certificate or certificate request, accordingly.
If set to none or if the option is not present then extensions
are ignored and not copied from SOURCE to TARGET.
If set to copy then any extension present in the SOURCE but not yet
present in TARGET is copied from SOURCE to TARGET.
If set to copyall then all extensions in the SOURCE are copied to
TARGET; if the extension is already present in TARGET (e.g. as a result
of -extfile option being specified), it is overwritten.

A new public function X509_to_X509_REQ_ex is also added. It differs
from the X509_to_X509_REQ function in that it takes flag as an
additional parameter, which determines whether it is necessary
to copy extensions from the certificate to the request.