ERR: Implement the macro ERR_raise() and use it
This macro uses a trick in C.  The following is permitted:

    #include <stdio.h>

    void first(void)
    {
        printf("Hello! ");
    }

    void foo(const char *bar)
    {
        printf("%s", bar);
    }

    int main()
    {
        /* This */
        (first(),foo)("cookie");
    }

ERR_raise() can be used to implement FUNCerr() as well, which takes
away the need for the special function ERR_put_func_error().

squash! ERR: Implement the macro ERR_raise() and use it

Make sure to mention ERR_raise_data() too